requirements.txt
faiss-cpu
openai
pypdf
requests
beautifulsoup4
tiktoken
langchain
chromadb